First try  speedup by reduce CoarseIsotpeGenerator
Reduce the calls to create a CoarseIsotopeGenerator by refactoring

Problem is that they don't require the exact same distrbution (one less
isotope) and we cannot just chop off the last isotope as this will throw
off the normalization.

Have the option to either renormalize (much more work) or just change
the test files so that they both require the same distribution.