[FEATURE] add output of nr spectra added
Add to OSW file FEATURE_MS2 table the NUM_SPEC_ADD column which the number of spectra added by the autoAdd.
If -Scoring:add_up_spectra is a positive number, all entries of will be
this positive number. If -Scoring:add_up_spectra is -1 than this column
will be variable based on peak width.